add time and duration to numeric tower
Time is implemented as a Number but no other
numbers can be converted to it.
It is parsed by 2018.12.23T02.07.00.
Times without date act as a duration and can be added to other times.
Internally durations are implemented as times < y1k.
They must have the same type as times, otherwise the elementary
operations would not work.
It implements comparison, dyadic - and +.