lui: add modes -a, -i
This adds non-gui modes to cmd/lui.
These are executed the same way as cmd/apl and cmd/iv, with the
only difference, that lui contains all libraries, while the
others only the minimum.

go test in lui also runs all tests in cmd/apl and cmd/iv
with the extended environment.

BUG: go test fails