channels: add channel-delay
Channel delay D<C is implemented to allow fixed rate animations.

Another feature would be a rated down monitor channel.
This would require a junction, but what to do with multiple loose ends?
Which one is the result of an expression? Can there be multiple?
Also: what to do, if several branches do concurrent writes?

A monitor channel could be used for log output during a long computation,
or a scope-like down rated image, while the main branch is not disturbed.

Syntax-wise this could be done as:
	Z←D<[f]C
Where the axis specification evaluates to a function (does the parser allow this?)
F would receive a channel with copied values as the right argument.

The more general syntax
	Z←<[f]C
could be used for a synchronised tee junction, or
	Z←<[f g h]C
for a fan-out pipe to 4 branches (the default + 3 extra).