roll: extend for negative and complex arguments
When called with negative arguments, roll returns a normally distributed
random number scaled by the argument.
When called with a complex argument, roll returns a random number with
a bivariat normal distributions, whose real and complex stdandard deviation
is equal to the real and imag part of the argument.

TODO: The random number generator is currently seeded with 1 when uset by Go.
This is ok for predictive simulations, but there should be an option to set the
seed explicitly.