add functional selective specification
Indexed assignment may contain a function withing an index specification.
This function is called for every item and receives the index vector on the left
and the scalar value on the right.
If the function returns true, the item is selected. If it returns false or empty
it is not selected.