value: change string method
The string method for apl.Value does not contain a pointer to the
complete apl state anymore, but a format value.

This also removed the dependency for apl state from object, table and
xgo At, Set methods.

Having less state is good for concurrent operations.
There was a race in iv/shape.iv test when using custom formats.
This should be fixed.